requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
cascade==0.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
default==0.1.2
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
Faker==10.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
imaplib2==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lark-parser==0.12.0
Markdown==3.3.4
MarkupSafe==2.0.1
mixer==7.2.0
packaging==21.3
password==0.2
progressbar==2.5
promise==2.3
psycopg2-binary==2.9.2
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
required==0.4.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Rx==1.6.1
self==2020.12.3
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
timestamp==0.0.1
uritemplate==4.1.1
urllib3==1.26.7
username==2.1.0
